code/google_docs_append/requirements.txt
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
